AT 01.01 | Freestyle project > View project page > Rename project Add testRenameProjectName() to FreestyleProject01Test.java Add RenameFreestyleProjectPage.java + inner methods Add some methods to FreestyleProjectPage https://trello.com/c/OKtQCE5L/1481-at-0101-freestyle-project-view-project-page-rename-project
