+++ src/test/java/school/redrover/FreestyleProject01Test.java
@@ -1,12 +1,9 @@
 package school.redrover;
 
 import org.testng.Assert;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 import school.redrover.model.FreestyleProjectPage;
 import school.redrover.model.MainPage;
-import school.redrover.model.NewJobPage;
 import school.redrover.runner.BaseTest;
 
 public class FreestyleProject01Test extends BaseTest {
@@ -74,26 +71,19 @@ public class FreestyleProject01Test extends BaseTest {
                 .selectBuildNow();
 
         Assert.assertEquals(freestyleProjectPage.getBuildsQuantity(), 1);
-
     }
 
-    @Test(dataProvider = "invalidSymbols")
-    public void testCreatingProjectWithInvalidName(String invalidSymbols){
-       String validationMessage = new MainPage(getDriver())
+    @Test
+    public void testRenameProjectName() {
+        FreestyleProjectPage freestyleProjectPage = new MainPage(getDriver())
                 .clickNewItem()
-                .selectFreestyleProject()
-                .enterItemName(invalidSymbols)
-                .getItemInvalidMessage();
+                .enterItemName(FREESTYLE_NAME)
+                .selectFreestyleProjectAndOk()
+                .clickSave()
+                .clickRenameProject()
+                .enterNewName(FREESTYLE_NAME + " New")
+                .submitNewName();
 
-       boolean okButton = new NewJobPage(getDriver())
-               .isOkButtonEnabled();
-
-        Assert.assertEquals(validationMessage, "» ‘" + invalidSymbols + "’ is an unsafe character");
-        Assert.assertFalse(okButton);
-    }
-
-    @DataProvider(name = "invalidSymbols")
-    public Object[][] invalidSymbols(){
-        return new Object[][] {{"!"},{"@"}, {"#"}, {"$"}, {"%"}, {"^"}, {"&"}, {"*"}, {":"}, {";"}, {"/"}, {"|"}, {"?"}, {"<"}, {">"}};
+        Assert.assertEquals(freestyleProjectPage.getProjectName(), "Project " + FREESTYLE_NAME + " New");
     }
 }
